Game_Error
Program giving wrong output as : 
"
I got no Idea what it mean "

Although the expected output is 

"You are in a dark room.
There is a door to your right and left.
Which one do you take?
> left
There is a bear here.
The bear has a bunch of honey.
The fat bear is in front of another door.
How are you going to move the bear?
> taunt bear
The bear has moved from the door. You can go through it now.
> open door
This room is full of gold. How much do you take?
> 1000 
You greedy bastard! Good job! "